% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_company.R
\name{all_company}
\alias{all_company}
\title{"Pull single company of data"}
\usage{
all_company()
}
\value{
a dataframe containing storm overflow data for
all water companies on the portal
}
\description{
"Pull single company of data"
}
\examples{
df <- all_company()
}
