% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_company.R
\name{single_company}
\alias{single_company}
\title{"Pull single company of data"}
\usage{
single_company(url)
}
\arguments{
\item{url}{for the API on waterUK portal}
}
\value{
a dataframe containing storm overflow data
}
\description{
"Pull single company of data"
}
\examples{
df <- single_company(url)
}
